refactor routes into testable functions & add test
closes #265 by searching for user based on username and email match

also, check for captcha sooner
