+++ db/models/__mocks__/index.js
@@ -1,0 +1,7 @@
+const Sequelize = require('sequelize');
+
+const db = {};
+
+db.Sequelize = Sequelize;
+
+module.exports = db;
